Add a docker workflow to the site publishing
Problem

It's a huge pain to try and maintain a consistent ruby environment on
ones development machine. There is a nonstop struggle to get the right
version of ruby, the gems installed, and everything coherent.

Solution

Use docker to build consistent environments. We add a Dockerfile which
bootstraps a Ubuntu 18.04 image with all the fixins and our entry point
becomes
```
$ bundle install
$ preview_the_site
$ publish_the_site
```